Helper Scripts/requirements.txt
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
